+++ src/Components/Form/index.tsx
@@ -1,5 +1,5 @@
 import { Dispatch, SetStateAction, createContext, useState } from "react";
-import FormField from "./FormField";
+import FormField from "../FormField";
 
 const initError = () => ({
   errorMessage: "",
