+++ src/Components/FormField/emotionCss.ts
@@ -1,0 +1,22 @@
+import { css } from "@emotion/react";
+
+export const titleCss = css({
+  fontSize: "18px",
+  fontWeight: "700",
+  lineHeight: "26px",
+  textAlign: "left",
+});
+
+export const descriptionCss = css({
+  fontSize: "10px",
+  fontWeight: "400",
+  lineHeight: "14px",
+  textAlign: "left",
+  color: "#8B95A1",
+  marginBottom: "30px",
+});
+
+export const rowStyle = css({
+  display: "flex",
+  justifyContent: "space-between",
+});
