+++ src/Components/Tooltip/emotionCss.ts
@@ -1,0 +1,8 @@
+import { css } from "@emotion/react";
+
+export const tooltipStyle = css({
+  color: "#D1180B",
+  marginTop: "10px",
+  fontSize: "0.8rem",
+  height: "1rem",
+});
