+++ src/Components/Tooltip/index.tsx
@@ -1,17 +1,10 @@
 /** @jsxImportSource @emotion/react */
-import { css } from "@emotion/react";
 import React from "react";
+import { tooltipStyle } from "./emotionCss";
 
 interface Props {
   children: React.ReactNode;
 }
-
-const tooltipStyle = css({
-  color: "#D1180B",
-  marginTop: "10px",
-  fontSize: "0.8rem",
-  height: "1rem",
-});
 
 const Tooltip = ({ children }: Props) => {
   return <div css={tooltipStyle}>{children}</div>;
